refactor: Move code to dtclient
